+++ CHANGELOG.md
@@ -9,12 +9,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.16.0] - 2018-02-01
+
 ### Added
 
 - gRPC Dial function
 - DB acceptance context
 - gRPC client acceptance context
 - App acceptance context
+
 
 ## [0.15.1] - 2018-01-24
 
@@ -263,7 +267,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bootstrap options
 
 
-[Unreleased]: https://github.com/goph/fxt/compare/v0.15.1...HEAD
+[Unreleased]: https://github.com/goph/fxt/compare/v0.16.0...HEAD
+[0.16.0]: https://github.com/goph/fxt/compare/v0.15.1...v0.16.0
 [0.15.1]: https://github.com/goph/fxt/compare/v0.15.0...v0.15.1
 [0.15.0]: https://github.com/goph/fxt/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/goph/fxt/compare/v0.13.0...v0.14.0